feat: Update external app URLs in PortalView
The external app URLs in the PortalView have been updated to use the correct URLs. This ensures that users are directed to the correct applications when they click on the corresponding links.

The "Gestion des Equipements" app URL has been changed to "http://169.239.64.130:8002/dashboard" and the "Gestion des Missions" app URL has been changed to "http://169.239.64.130:8002/missions". Additionally, the "Gestion des Immobilisations" app URL has been updated to "http://immobilisations.arti.ci".

This commit also includes a change in the active profile in the application.properties file, setting it to "prod".

These changes were made to ensure that the application functions correctly and directs users to the appropriate external applications.

```